Merged PR 1260: Merge feature/4470-4969-remove-unsecure-mcf-connections to master
Changed all references to config.McfEndpoint to config.SecureMcfEndpoint
Config settings are in a NuGet package so that won't be changed
Also supressed warning 1591.  This generates hundreds of warnings that mask the handful that may be significant.
Also downgraded a recent NuGet Core update that was causing some bugs.  Will track down in another commit.
Related work items: #4969